+++ toolchain/tests/programs/Cargo.toml
@@ -7,17 +7,17 @@ edition = "2021"
 [dependencies]
 axvm = { path = "../../riscv/axvm" }
 axvm-platform = { path = "../../riscv/platform" }
+axvm-ecc = { path = "../../../lib/ecc", default-features = false }
 serde = { version = "1.0", default-features = false, features = [
     "alloc",
     "derive",
 ] }
 hex = { version = "0.4.3", default-features = false, features = ["alloc"] }
 hex-literal = { version = "0.4.1", default-features = false }
-axvm-ecc = { path = "../../../lib/ecc", default-features = false }
 
 [features]
 default = []
-std = ["axvm/std"]
+std = ["axvm/std", "axvm-ecc/std"]
 
 [profile.release]
 panic = "abort"
