+++ toolchain/tests/src/ecc_tests.rs
@@ -27,3 +27,23 @@ fn test_moduli_setup_runtime() -> Result<()> {
     assert!(!executor.config.supported_modulus.is_empty());
     Ok(())
 }
+
+#[test]
+fn test_modular_runtime() -> Result<()> {
+    let elf = build_example_program("little")?;
+    let executor = VmExecutor::<F>::new(VmConfig::rv32im().add_canonical_modulus());
+    executor.execute(elf, vec![])?;
+    Ok(())
+}
+
+#[test]
+fn test_ec_runtime() -> Result<()> {
+    let elf = build_example_program("ec")?;
+    let executor = VmExecutor::<F>::new(
+        VmConfig::rv32im()
+            .add_canonical_modulus()
+            .add_canonical_ec_curves(),
+    );
+    executor.execute(elf, vec![])?;
+    Ok(())
+}
