+++ toolchain/tests/src/lib.rs
@@ -4,6 +4,5 @@ pub mod utils;
 
 #[cfg(test)]
 pub mod basic_tests;
-
 #[cfg(test)]
-pub mod moduli_setup_test;
+pub mod ecc_tests;
